【フロント・機能】経費一覧（西尾）

+++ demo2309/src/main/java/com/example/demo/repository/ExpenseListRepository.java
@@ -9,4 +9,6 @@ import com.example.demo.entity.ExpenseListEntity;
  * ユーザー情報 Repository
  */
 @Repository
-public interface ExpenseListRepository extends JpaRepository<ExpenseListEntity, Long> {}
+public interface ExpenseListRepository extends JpaRepository<ExpenseListEntity, Integer> {
+	
+}
